Moved unit tests
