+++ tests/Doctrine/Tests/Models/VersionedOneToOne/SecondRelatedEntity.php
